compiler: Move expose_tuning_knobs to utils
